+++ FluentSecurity/Policy/RequireRolePolicy.cs
@@ -25,10 +25,11 @@ namespace FluentSecurity.Policy
 			if (context.CurrenUserAuthenticated() == false)
 				return PolicyResult.CreateFailureResult(this, "Anonymous access denied");
 
-			if (context.CurrenUserRoles() == null || context.CurrenUserRoles().Any() == false)
+			var currentUserRoles = context.CurrenUserRoles().EnsureIsList();
+			if (currentUserRoles.Any() == false)
 				return PolicyResult.CreateFailureResult(this, "Access denied");
 
-			if (context.CurrenUserRoles().Any(role => _requiredRoles.Contains(role)) == false)
+			if (currentUserRoles.Any(role => _requiredRoles.Contains(role)) == false)
 			{
 				const string message = "Access requires one of the following roles: {0}.";
 				var formattedMessage = string.Format(message, GetRoles());